Migrarion des methodes de PublicController dans MessageController
