+++ views/message/allMessage.php
@@ -5,7 +5,7 @@ ob_start();
 <div class="container">
     <div class="d-flex justify-content-between align-items-center">
         <div class="fs-1">Liste des messages</div>
-        <a href="/new">
+        <a href="/message/new">
             <button type="button" class="btn btn-primary">Nouvelle message</button>
         </a>
     </div>
